refactor(fetch-token-mint-info): handle devnet dummy token and move fetch token file
