+++ src/api/otc-state/fetchOtcState.ts
@@ -12,7 +12,9 @@ import { RedeemLogicForwardState } from 'models/RedeemLogicForwardState';
 
 import PROGRAMS from '../../configs/programs.json';
 import { OtcState } from '../../models/OtcState';
-import { fetchTokenInfo } from './fetchTokenInfo';
+import { fetchTokenInfo } from '../tokens/fetchTokenInfo';
+
+const [DUMMY_TOKEN_MINT, USDC_MINT] = ['7XSvJnS19TodrQJSbjUR6tEGwmYyL1i9FX7Z5ZQHc53W', 'EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v'];
 
 export const fetchOtcState = async (provider: AnchorProvider, otcStateAddress: PublicKey): Promise<OtcState> => {
 	const promises: Promise<void>[] = [];
@@ -26,7 +28,12 @@ export const fetchOtcState = async (provider: AnchorProvider, otcStateAddress: P
 	const res = new OtcState();
 	res.publickey = otcStateAddress;
 	res.reserveMintInfo = await getMint(provider.connection, trancheConfigAccountInfo.reserveMint);
-	res.reserveTokenInfo = await fetchTokenInfo(provider.connection, trancheConfigAccountInfo.reserveMint);
+
+	if (trancheConfigAccountInfo.reserveMint.toBase58() === DUMMY_TOKEN_MINT) {
+		res.reserveTokenInfo = await fetchTokenInfo(provider.connection, new PublicKey(USDC_MINT));
+	} else {
+		res.reserveTokenInfo = await fetchTokenInfo(provider.connection, trancheConfigAccountInfo.reserveMint);
+	}
 
 	res.createdAt = accountInfo.created.toNumber() * 1000;
 	res.depositAvailableFrom = accountInfo.depositStart.toNumber() * 1000;
